+++ srcCpp/perftest_publisher.cxx
@@ -392,18 +392,18 @@ bool perftest_cpp::ParseConfig(int argc, char *argv[])
 
             _DataLen = strtol(argv[i], NULL, 10);
 
-            if (_DataLen < (unsigned)OVERHEAD_BYTES)
+            if (_DataLen < (unsigned long)OVERHEAD_BYTES)
             {
                 fprintf(stderr, "-dataLen must be >= %d\n", OVERHEAD_BYTES);
                 return false;
             }
-            if (_DataLen > (unsigned)MAX_PERFTEST_SAMPLE_SIZE)
+            if (_DataLen > (unsigned long)MAX_PERFTEST_SAMPLE_SIZE)
             {
                 fprintf(stderr,"-dataLen must be <= %d\n", MAX_PERFTEST_SAMPLE_SIZE);
                 return false;
             }
 
-            if (_useUnbounded == 0 && _DataLen > (unsigned)MAX_BOUNDED_SEQ_SIZE) {
+            if (_useUnbounded == 0 && _DataLen > (unsigned long)MAX_BOUNDED_SEQ_SIZE) {
                 _useUnbounded = MAX_BOUNDED_SEQ_SIZE;
             }
         }
@@ -433,12 +433,12 @@ bool perftest_cpp::ParseConfig(int argc, char *argv[])
 
                 _useUnbounded = strtol(argv[i], NULL, 10);
 
-                if (_useUnbounded < (unsigned)OVERHEAD_BYTES)
+                if (_useUnbounded < (unsigned long)OVERHEAD_BYTES)
                 {
                     fprintf(stderr, "-unbounded <managerMemory> must be >= %d\n", OVERHEAD_BYTES);
                     return false;
                 }
-                if (_useUnbounded > (unsigned)MAX_PERFTEST_SAMPLE_SIZE)
+                if (_useUnbounded > (unsigned long)MAX_PERFTEST_SAMPLE_SIZE)
                 {
                     fprintf(stderr,"-unbounded <managerMemory> must be <= %d\n", MAX_PERFTEST_SAMPLE_SIZE);
                     return false;
