+++ srcCpp03/RTIDDSImpl.cxx
@@ -249,18 +249,18 @@ bool RTIDDSImpl<T>::ParseConfig(int argc, char *argv[])
 
             _DataLen = strtol(argv[i], NULL, 10);
 
-            if (_DataLen < (unsigned)perftest_cpp::OVERHEAD_BYTES) {
+            if (_DataLen < (unsigned long)perftest_cpp::OVERHEAD_BYTES) {
                 std::cerr << "[Error] -dataLen must be >= "
                         << perftest_cpp::OVERHEAD_BYTES << std::endl;
                 throw std::logic_error("[Error] Error parsing commands");
             }
 
-            if (_DataLen > (unsigned)MAX_PERFTEST_SAMPLE_SIZE) {
+            if (_DataLen > (unsigned long)MAX_PERFTEST_SAMPLE_SIZE) {
                 std::cerr << "[Error] -dataLen must be <= "
                         << MAX_PERFTEST_SAMPLE_SIZE << std::endl;
                 throw std::logic_error("[Error] Error parsing commands");
             }
-            if (_useUnbounded == 0 && _DataLen > (unsigned)MAX_BOUNDED_SEQ_SIZE) {
+            if (_useUnbounded == 0 && _DataLen > (unsigned long)MAX_BOUNDED_SEQ_SIZE) {
                 _useUnbounded = MAX_BOUNDED_SEQ_SIZE;
             }
         }
@@ -272,13 +272,13 @@ bool RTIDDSImpl<T>::ParseConfig(int argc, char *argv[])
                 ++i;
                 _useUnbounded = strtol(argv[i], NULL, 10);
 
-                if (_useUnbounded <  (unsigned)perftest_cpp::OVERHEAD_BYTES)
+                if (_useUnbounded <  (unsigned long)perftest_cpp::OVERHEAD_BYTES)
                 {
                     std::cerr << "[Error] -unbounded <managerMemory> must be >="
                         << perftest_cpp::OVERHEAD_BYTES << std::endl;
                     throw std::logic_error("[Error] Error parsing commands");
                 }
-                if (_useUnbounded > (unsigned)MAX_PERFTEST_SAMPLE_SIZE)
+                if (_useUnbounded > (unsigned long)MAX_PERFTEST_SAMPLE_SIZE)
                 {
                     std::cerr << "[Error] -unbounded <managerMemory> must be <="
                         << MAX_PERFTEST_SAMPLE_SIZE << std::endl;
@@ -668,7 +668,7 @@ bool RTIDDSImpl<T>::ParseConfig(int argc, char *argv[])
         }
     }
 
-    if (_DataLen > (unsigned)MAX_SYNCHRONOUS_SIZE) {
+    if (_DataLen > (unsigned long)MAX_SYNCHRONOUS_SIZE) {
         if (_isScan) {
             std::cerr << "[Info] DataLen will be ignored since -scan is present."
                     << std::endl;
@@ -695,7 +695,7 @@ bool RTIDDSImpl<T>::ParseConfig(int argc, char *argv[])
      * nor if the sample is bigger (in this case we avoid the checking in the
      * middleware).
      */
-     if (_BatchSize > 0 && (unsigned)_BatchSize <= _DataLen) {
+     if (_BatchSize > 0 && (unsigned long)_BatchSize <= _DataLen) {
          std::cerr << "[Info] Batching dissabled: BatchSize (" << _BatchSize
                    << ") is equal or smaller than the sample size (" << _DataLen
                    << ")."  << std::endl;
